fixes: folders refactoring
