+++ src/App.tsx
@@ -3,15 +3,14 @@ import {
   RouterProvider, redirect
 } from "react-router-dom";
 import Home from "./pages/Home";
-import Todo from "./pages/TodoCalendar";
 import Landing from "./pages/Landing";
 import './App.css';
 import MainLayout from "./layouts/MainLayout";
 import AuthLayout from "./layouts/AuthLayout";
-import RegisterForm from "./components/RegisterForm";
-import LoginForm from "./components/LoginForm";
+import RegisterForm from "./components/auth/RegisterForm";
+import LoginForm from "./components/auth/LoginForm";
 import LandingLayout from "./layouts/LandingLayout";
-import TodoCalendar from "./components/TodoList";
+import TodoCalendar from "./components/todo/TodoList";
 
 function checkAuth() {
   const user = localStorage.getItem("usuario");
