+++ src/api/deleteTodo.ts
@@ -1,4 +1,4 @@
-import supabase from "../supabaseClient";
+import supabase from "../lib/supabaseClient";
 
 export async function deleteTodoItem(id: number) {
   return supabase
