+++ src/api/login.ts
@@ -1,4 +1,4 @@
-import supabase from "../supabaseClient";
+import supabase from "../lib/supabaseClient";
 
 export async function loginUser(email: string, password: string) {
