+++ src/components/auth/LoginForm.tsx
@@ -1,5 +1,5 @@
 import { NavLink, useNavigate } from "react-router-dom";
-import { loginUser } from "../services/login";
+import { loginUser } from "../../api/login";
 import { useState } from "react";
 import { IconArrowBarLeft} from "@tabler/icons-react"
 
@@ -8,7 +8,7 @@ function Login() {
 
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
+  const [error, setError] = useState<string | null>(null);
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
