+++ src/components/todo/TodoForm.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import supabase from "../supabaseClient";
+import supabase from "../../lib/supabaseClient";
 import { DatePicker, Modal, Button } from "antd";
 import dayjs from "dayjs";
 import Swal from "sweetalert2";
