+++ src/lib/supabaseClient.ts
@@ -1,5 +1,5 @@
 import { createClient } from '@supabase/supabase-js'
-import type { Database } from '../database.types'
+import type { Database } from '../../database.types'
 
 const SUPABASE_URL = import.meta.env.VITE_SUPABASE_URL
 const SUPABASE_API_KEY = import.meta.env.VITE_SUPABASE_API_KEY
